project/text-to-sql-prototype/backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
asyncpg==0.29.0
aiomysql==0.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-dotenv==1.0.0

# Utilities
loguru==0.7.2
httpx==0.26.0
tenacity==8.2.3

# Development
pytest==7.4.4
pytest-asyncio==0.21.1
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
